Merged PR 2725: SEB-1476 Allow users to upload images for resale certs
## Description

This allows users to upload images and pdf files as resale certs.
When an image is selected a blank pdf file is created with javascript and the image is added to that file. Then that file is uploaded like a regular pdf that was selected would be.

For Reference # SEB-1476

## Type of change
- [x] New feature (non-breaking change which adds functionality)

## Checklist:
- [x] I have performed a self-review of my own code