Merged PR 2737: SEB-1499 bug in automatically applied promotions
## Description

When automatically applied promotions the function checks if the order is valid, then removes all promotions if the order is not valid. However there may be cases when the order is not valid but there are also no promotions applied.
This update does not remove promotions if there are none applied.

For Reference # [SEB-1499]

## Type of change
- [x] Bug fix (non-breaking change which fixes an issue)

## Checklist:
- [x] I have performed a self-review of my own code

SEB-1499